requirements.txt

#This file standardizes the environment setup so every developer and CI agent installs the same dependencies
pytest-bdd==8.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-playwright==0.7.2
pytest-xdist==3.8.0
python-dotenv==1.2.2
python-slugify==8.0.4
referencing==0.37.0
requests==2.32.5
requests-mock==1.12.1
rpds-py==0.30.0
six==1.17.0
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3